cleanup(netx): merge integration_test.go into netx_test.go
See https://github.com/ooni/probe/issues/2121
